+++ src/js/constants/objects/cardsList.js
@@ -1,0 +1,7 @@
+import { cardsListElement } from './elements';
+import CardsList from '../../../blocks/cards-list/CardsList';
+import addCardTrash from '../../utilits/callbacks/cardsList';
+
+const cardsList = new CardsList(cardsListElement, addCardTrash);
+
+export default cardsList;
