+++ src/js/constants/objects/header.js
@@ -9,8 +9,7 @@ import {
 
 import Header from '../../../blocks/header/Header';
 import { getProfile } from '../../utilits/functions';
-import { logoutHandlerClick } from '../../utilits/callbacks';
-
+import logoutHandlerClick from '../../utilits/callbacks';
 
 /** Объект хёдера, в основном управляет всей логикой приложения.
   * Инициализирует начальную страницу.
