+++ src/js/constants/objects/searchForm.js
@@ -1,7 +1,7 @@
 import { searchFormElement } from './elements';
 
 import SearchForm from '../../../blocks/search/__form/SearchForm';
-import { searchNews } from '../../utilits/callbacks';
+import searchNews from '../../utilits/callbacks/searchForm';
 
 /** Объект формы поиска новостей */
 const searchForm = new SearchForm(searchFormElement, searchNews);
