+++ src/saved-articles/articles.js
@@ -2,21 +2,19 @@ import '../pages/articles.css';
 
 import { profileOwner, NOT_CREATE_RESOURCE, NOT_ARTICLES } from '../js/constants/constants';
 import { usersApi } from '../js/constants/objects/api';
-import { cardsListElement, userArticlesContainer } from '../js/constants/objects/elements';
+import { userArticlesContainer } from '../js/constants/objects/elements';
 import header from '../js/constants/objects/header';
+import cardsList from '../js/constants/objects/cardsList';
+
 import articlesIntro from '../js/constants/objects/articlesIntro';
 
-import CardsList from '../blocks/cards-list/CardsList';
 import { getProfile, translateUsersApiParametrsToCardParametrs } from '../js/utilits/functions';
 import {
   renderLoginHeaderArticles,
   renderNotLoginHeaderArticles,
-  addCardTrash,
-} from '../js/utilits/callbacks';
+} from '../js/utilits/callbacks/header';
 
 function main() {
-  const cardsList = new CardsList(cardsListElement, addCardTrash);
-
   header.create(renderLoginHeaderArticles, renderNotLoginHeaderArticles);
   articlesIntro.create();
   articlesIntro.render();
